add some test
