+++ lib/utils.js
@@ -1650,3 +1650,29 @@ export function stringToObject(prefilledValues) {
   })
   return outputObject
 }
+
+/**
+ * checks whether email has instituion domain or subdomain
+ *
+ * @param {string} email - the email address to check
+ * @param {Array} institutionDomains - the list of institution domains
+ * @returns boolean
+ */
+export function isInstitutionEmail(email, institutionDomains) {
+  const emailDomainTokens = email.split('@').pop()?.trim()?.toLowerCase().split('.')
+  return institutionDomains.some((institutionDomain) => {
+    const institutionDomainTokens = institutionDomain.split('.')
+    if (institutionDomainTokens.length > emailDomainTokens.length) {
+      return false
+    }
+    for (let i = 1; i <= institutionDomainTokens.length; i += 1) {
+      if (
+        institutionDomainTokens[institutionDomainTokens.length - i] !==
+        emailDomainTokens[emailDomainTokens.length - i]
+      ) {
+        return false
+      }
+    }
+    return true
+  })
+}
